+++ src/addons/addons/ai-integration/helpers/attachment.js
