+++ src/addons/addons/ai-integration/helpers/parser.js
